+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Content Security Policy for configured Matomo servers in backend for TYPO3 v12 (#45)
 
+### Fixed
+- Superfluous chart.css file is not embedded anymore in TYPO3 v12
+- ESM modules instead of RequireJS is used in TYPO3 v12
+
 ## [2.1.1] - 2023-04-01
 
 ### Fixed