+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.2.0] - 2023-05-21
+
 ### Added
 - Content Security Policy for configured Matomo servers in backend for TYPO3 v12 (#45)
 
@@ -159,7 +161,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial release
 
 
-[Unreleased]: https://github.com/brotkrueml/typo3-matomo-widgets/compare/v2.1.1...HEAD
+[Unreleased]: https://github.com/brotkrueml/typo3-matomo-widgets/compare/v2.2.0...HEAD
+[2.2.0]: https://github.com/brotkrueml/typo3-matomo-widgets/compare/v2.1.1...v2.2.0
 [2.1.1]: https://github.com/brotkrueml/typo3-matomo-widgets/compare/v2.1.0...v2.1.1
 [2.1.0]: https://github.com/brotkrueml/typo3-matomo-widgets/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/brotkrueml/typo3-matomo-widgets/compare/v1.7.0...v2.0.0